+++ src/states/chatting.ts
@@ -1,0 +1,17 @@
+import { atom } from "recoil";
+import * as Stomp from "stompjs";
+
+export const stompClientState = atom<Stomp.Client | null>({
+  key: "stompClientState",
+  default: null,
+});
+
+export const messageState = atom<string[]>({
+  key: "messageState",
+  default: [],
+});
+
+export const activeRoomIdState = atom<number>({
+  key: "activeRoomIdState",
+  default: 1,
+});
